Split BotorchRecommender into submodules
