+++ baybe/recommenders/pure/bayesian/botorch/__init__.py
@@ -1,0 +1,7 @@
+"""Botorch recommender."""
+
+from baybe.recommenders.pure.bayesian.botorch.core import BotorchRecommender
+
+__all__ = [
+    "BotorchRecommender",
+]
